Fix declaration redundancy

See merge request xdol/go-yaml!13

+++ path_test.go
@@ -97,12 +97,12 @@ store:
 		{
 			name:     "$.store.bicycle.price",
 			path:     builder().Root().Child("store").Child("bicycle").Child("price").Build(),
-			expected: float64(19.95),
+			expected: 19.95,
 		},
 		{
 			name:     `$.store.'bicycle*unicycle'.price`,
 			path:     builder().Root().Child("store").Child(`bicycle*unicycle`).Child("price").Build(),
-			expected: float64(20.25),
+			expected: 20.25,
 		},
 	}
 	t.Run("PathString", func(t *testing.T) {
